Switch from individual routes to handlers.
