+++ cmd/web/main.go
@@ -9,7 +9,7 @@ import (
 	"os/signal"
 	"time"
 
-	"github.com/mikestefanello/pagoda/pkg/routes"
+	"github.com/mikestefanello/pagoda/pkg/handlers"
 	"github.com/mikestefanello/pagoda/pkg/services"
 )
 
@@ -23,7 +23,7 @@ func main() {
 	}()
 
 	// Build the router
-	routes.BuildRouter(c)
+	handlers.BuildRouter(c)
 
 	// Start the server
 	go func() {
