+++ pkg/handlers/error.go
@@ -1,4 +1,4 @@
-package routes
+package handlers
 
 import (
 	"net/http"
@@ -9,11 +9,11 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-type errorHandler struct {
+type Error struct {
 	controller.Controller
 }
 
-func (e *errorHandler) Get(err error, ctx echo.Context) {
+func (e *Error) Page(err error, ctx echo.Context) {
 	if ctx.Response().Committed || context.IsCanceledError(err) {
 		return
 	}
