+++ pkg/handlers/handlers_test.go
@@ -1,4 +1,4 @@
-package routes
+package handlers
 
 import (
 	"net/http"
