+++ pkg/handlers/pages.go
@@ -1,6 +1,7 @@
-package routes
+package handlers
 
 import (
+	"fmt"
 	"html/template"
 
 	"github.com/mikestefanello/pagoda/pkg/controller"
@@ -8,9 +9,19 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
+const (
+	routeNameAbout = "about"
+	routeNameHome  = "home"
+)
+
 type (
-	about struct {
+	Pages struct {
 		controller.Controller
+	}
+
+	post struct {
+		Title string
+		Body  string
 	}
 
 	aboutData struct {
@@ -25,7 +36,38 @@ type (
 	}
 )
 
-func (c *about) Get(ctx echo.Context) error {
+func (c *Pages) Routes(g *echo.Group) {
+	g.GET("/", c.Home).Name = routeNameHome
+	g.GET("/about", c.About).Name = routeNameAbout
+}
+
+func (c *Pages) Home(ctx echo.Context) error {
+	page := controller.NewPage(ctx)
+	page.Layout = "main"
+	page.Name = "home"
+	page.Metatags.Description = "Welcome to the homepage."
+	page.Metatags.Keywords = []string{"Go", "MVC", "Web", "Software"}
+	page.Pager = controller.NewPager(ctx, 4)
+	page.Data = c.fetchPosts(&page.Pager)
+
+	return c.RenderPage(ctx, page)
+}
+
+// fetchPosts is an mock example of fetching posts to illustrate how paging works
+func (c *Pages) fetchPosts(pager *controller.Pager) []post {
+	pager.SetItems(20)
+	posts := make([]post, 20)
+
+	for k := range posts {
+		posts[k] = post{
+			Title: fmt.Sprintf("Post example #%d", k+1),
+			Body:  fmt.Sprintf("Lorem ipsum example #%d ddolor sit amet, consectetur adipiscing elit. Nam elementum vulputate tristique.", k+1),
+		}
+	}
+	return posts[pager.GetOffset() : pager.GetOffset()+pager.ItemsPerPage]
+}
+
+func (c *Pages) About(ctx echo.Context) error {
 	page := controller.NewPage(ctx)
 	page.Layout = "main"
 	page.Name = "about"
