--- pkg/routes/logout.go
@@ -1,21 +1,0 @@
-package routes
-
-import (
-	"github.com/mikestefanello/pagoda/pkg/controller"
-	"github.com/mikestefanello/pagoda/pkg/msg"
-
-	"github.com/labstack/echo/v4"
-)
-
-type logout struct {
-	controller.Controller
-}
-
-func (l *logout) Get(c echo.Context) error {
-	if err := l.Container.Auth.Logout(c); err == nil {
-		msg.Success(c, "You have been logged out successfully.")
-	} else {
-		msg.Danger(c, "An error occurred. Please try again.")
-	}
-	return l.Redirect(c, "home")
-}
